+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/src/main/resources/com/liferay/portal/workflow/metrics/rest/internal/resource/helper/dependencies/workflow-metrics-instance-count-init-script.painless
@@ -1,5 +1,5 @@
-state.assigneeUserIdInstanceIds = new HashSet();
-state.assigneeUserIds = params.assigneeUserIds;
+state.assigneeIdInstanceIds = new HashSet();
+state.assigneeIds = params.assigneeIds;
 state.endDate = params.endDate;
 state.slaStatuses = params.slaStatuses;
 state.slaStatusMap = new HashMap();
